fix for omnipath: changing gaspi_notification_id from unsigned int to byte length, because omnipath writes single bytes at a time, which can result in races because (a) unexpected notification value is read out (b) zero value after reset is filled, bug fix: restrict the notification value to byte length which can not be overwritten: fix has been tested on Omnipath system at kraken.cerfacs.fr

+++ include/GASPI.h
@@ -29,6 +29,8 @@ along with GPI-2. If not, see <http://www.gnu.org/licenses/>.
 #ifndef GPI2_H
 #define GPI2_H
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C"
 {
@@ -61,7 +63,7 @@ extern "C"
   typedef float gaspi_time_t;
   typedef unsigned long gaspi_cycles_t;
   typedef unsigned short gaspi_notification_id_t;
-  typedef unsigned int gaspi_notification_t;
+  typedef uint8_t gaspi_notification_t;
   typedef unsigned int gaspi_statistic_counter_t;
   typedef char * gaspi_string_t;
